Move test files, add newlines
