+++ src/tests/store/users.test.js
@@ -12,4 +12,15 @@ describe('Store', () => {
       loggedInUser: {}
     })
   })
+
+  it('should update after signup', async () => {
+    await store.dispatch({
+      type: 'POST_SIGNUP_INFO',
+      payload: ['Rerun them']
+    })
+
+    expect(store.getState()).toEqual({
+      users: ['Rerun them']
+    })
+  })
 })
